redesigned lexer and parser
