+++ index.js
@@ -51,7 +51,7 @@ Client.db.getConnection()
             connection = await checkConnection(Client.db, connection);
             let env = await createUserTermEnv(Client, connection, message);
             let script = prepareScript(env, message.content);
-            if(script.length)spawnProcess(env, null, script[0].toLowerCase(), script);
+            if(script.length)spawnProcess(env, null, script[0].toLowerCase(), script, [], []);
             if(isUserMention(message.content) && getUserID(message.content) === Client.user.id)message.reply(languages.get(env.serverConfig.getLanguage()).help_dialog.replace("$prefix", env.serverConfig.getPrefix()).replace("$prefix", env.serverConfig.getPrefix()));
         });
 
@@ -61,7 +61,7 @@ Client.db.getConnection()
                 connection = await checkConnection(Client.db, connection);
                 let env = await createUserTermEnv(Client, connection, newMessage);
                 let script = prepareScript(env, newMessage.content);
-                if(script.length)spawnProcess(env, null, script[0].toLowerCase(), script);
+                if(script.length)spawnProcess(env, null, script[0].toLowerCase(), script, [], []);
                 if(isUserMention(newMessage.content) && getUserID(newMessage.content) === Client.user.id)newMessage.reply(languages.get(env.serverConfig.getLanguage()).help_dialog.replace("$prefix", env.serverConfig.getPrefix()).replace("$prefix", env.serverConfig.getPrefix()));
             }
         })
